fix bug where study would be invalid
The problem was that when there were no studies saved and the code tried
to validate a new study. user.studies would also return the study that
is has not yet been saved. This would cause a condition to fail thus
making the study invalid.

This has now been fixed by first rejecting the studies that are not
saved.