Refactor ExportsController by changing what the Export object does
Previously it would just be a dump model like object that held onto
techniques, categories and belts. It didn't have any behavior.

I changed that so Export is no longer a model, but instead a service.
This service can then filter the techniques its given to only include
those that are in certain categories and belts.